+++ FrontEnd/lib/Enter Number Page.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:tap_on/Enter%20verification%20page.dart';
 import 'package:tap_on/constants.dart';
 
-class Enternumber extends StatelessWidget{
+class Enternumber extends StatelessWidget {
   const Enternumber({super.key});
 
-@override
+  @override
   Widget build(BuildContext context) {
     double height = MediaQuery.of(context).size.height;
     double width = MediaQuery.of(context).size.width;
@@ -62,10 +62,15 @@ class Enternumber extends StatelessWidget{
                       ),
                     ),
                   ),
-                  const SizedBox(height: 25,),
+                  const SizedBox(
+                    height: 25,
+                  ),
                   ElevatedButton(
                     onPressed: () {
-                      Navigator.push(context, MaterialPageRoute(builder:(context) => VerificationScreen()));
+                      Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) => VerificationScreen()));
                       // Add your onPressed logic here
                     },
                     style: ElevatedButton.styleFrom(
@@ -91,17 +96,5 @@ class Enternumber extends StatelessWidget{
         ],
       ),
     );
-
-
-
-
-
-
-
   }
-
-
-
-
-
 }
