+++ FrontEnd/lib/Providers/providerFeedback.dart
@@ -1,12 +1,6 @@
 import 'package:flutter/material.dart';
 
-void main() {
-  runApp(MaterialApp(
-    home: ReviewsPage(),
-  ));
-}
-
-class ReviewsPage extends StatelessWidget {
+class PrviderFeedbackPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -48,7 +42,7 @@ class ReviewsPage extends StatelessWidget {
               child: ListView(
                 children: [
                   _buildReview(
-                    'Joan Perkins',
+                    'Mohamme Rishaf',
                     'This chair is a great addition for any room in your home, not only just the living room. Featuring a mid-century design with modern available on the market. However, and with that said, if you are like most people in the market, it is just perfect!',
                     5,
                     '1 days ago',
@@ -70,7 +64,6 @@ class ReviewsPage extends StatelessWidget {
             ),
 
             // Write a review button
-            
           ],
         ),
       ),
@@ -90,7 +83,8 @@ class ReviewsPage extends StatelessWidget {
     );
   }
 
-  Widget _buildRatingRow(String label, int stars, Color color, double widthFactor) {
+  Widget _buildRatingRow(
+      String label, int stars, Color color, double widthFactor) {
     return Row(
       children: [
         Text(label),
