+++ FrontEnd/lib/ShopOwner/ShopNotfication.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:tap_on/Home%20page.dart';
-import 'package:tap_on/providerDashboard.dart';
+import 'package:tap_on/ShopOwner/ShopOwnerDashboard.dart';
 
-class ProviderNotificationPage extends StatelessWidget {
-  const ProviderNotificationPage({super.key});
+class Shopnotfication extends StatelessWidget {
+  const Shopnotfication({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -16,30 +15,29 @@ class ProviderNotificationPage extends StatelessWidget {
           leading: IconButton(
             icon: const Icon(Icons.arrow_back),
             onPressed: () {
-              Navigator.push(
-                  context, MaterialPageRoute(builder: (context) => Providerdashboard()));
+              Navigator.push(context,
+                  MaterialPageRoute(builder: (context) => ShopdashboardPage()));
             },
           ),
-
         ),
         body: ListView(
           padding: const EdgeInsets.all(16.0),
           children: [
             NotificationTile(
               title: "Meeting reminder",
-              subtitle: "Today at 2:00 PM",
+              subtitle: "TYou have prebooking order today at 2:00 PM",
             ),
             NotificationTile(
-              title: "Your Service Accepted",
+              title: "You have feedback",
+              subtitle: "Your equipments are good",
+            ),
+            NotificationTile(
+              title: "Your item confirmed",
               subtitle: "Track order status",
             ),
             NotificationTile(
-              title: "Your Service Added Succesfully",
-              subtitle: "Track service status",
-            ),
-            NotificationTile(
               title: "Up to 50% off",
-              subtitle: "Track service status",
+              subtitle: "Track order status",
             ),
           ],
         ),
@@ -52,7 +50,8 @@ class NotificationTile extends StatelessWidget {
   final String title;
   final String subtitle;
 
-  const NotificationTile({super.key, required this.title, required this.subtitle});
+  const NotificationTile(
+      {super.key, required this.title, required this.subtitle});
 
   @override
   Widget build(BuildContext context) {
