+++ FrontEnd/lib/ShopOwner/ShopOwnerDashboard.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:tap_on/AddService.dart';
 import 'package:tap_on/Home%20page.dart';
-import 'package:tap_on/providerFeedback.dart';
-import 'package:tap_on/providerNotification.dart';
-import 'package:tap_on/providerOrderHistory.dart';
-import 'package:tap_on/providerProfile.dart';
-import 'package:tap_on/servicemanage.dart';
+import 'package:tap_on/ShopOwner/ShopNotfication.dart';
+import 'package:tap_on/ShopOwner/ShopOrderhistory.dart';
+import 'package:tap_on/ShopOwner/AddTools.dart';
+import 'package:tap_on/ShopOwner/ShopProfile.dart';
+import 'package:tap_on/ShopOwner/shopfeedback.dart';
 
+import 'package:tap_on/ShopOwner/ToolsManagement.dart';
 
-class Providerdashboard extends StatelessWidget {
+class ShopdashboardPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -18,7 +18,7 @@ class Providerdashboard extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             const Text(
-              'TapOn Provider',
+              'TapOn ShopOwner',
               style: TextStyle(
                 fontSize: 20, // You can adjust the size
                 fontWeight: FontWeight.bold, // Optional: Makes the text bold
@@ -72,7 +72,7 @@ class Providerdashboard extends StatelessWidget {
                 Navigator.push(
                     context,
                     MaterialPageRoute(
-                        builder: (context) => Providerorderhistory()));
+                        builder: (context) => ShopOrderhistoryPage()));
                 // Handle Orders button press
               },
             ),
@@ -80,23 +80,23 @@ class Providerdashboard extends StatelessWidget {
             // Menu Manager button
             ListTile(
               leading: Icon(Icons.add),
-              title: Text('Add Service'),
+              title: Text('Add Tool'),
               onTap: () {
                 Navigator.push(
                     context,
                     MaterialPageRoute(
                         builder: (context) =>
-                            AddServicePage())); // Handle Menu Manager button press
+                            AddItemPage())); // Handle Menu Manager button press
               },
             ),
 
             // Performance button
             ListTile(
               leading: Icon(Icons.edit),
-              title: Text('Service Manager'),
+              title: Text('Menu Manager'),
               onTap: () {
                 Navigator.push(context,
-                    MaterialPageRoute(builder: (context) => ManageService()));
+                    MaterialPageRoute(builder: (context) => MenuScreen()));
                 // Handle Performance button press
               },
             ),
@@ -107,7 +107,7 @@ class Providerdashboard extends StatelessWidget {
               title: Text('Notifications'),
               onTap: () {
                 Navigator.push(context,
-                    MaterialPageRoute(builder: (context) => ProviderNotificationPage()));
+                    MaterialPageRoute(builder: (context) => Shopnotfication()));
 
                 // Handle Notifications button press
               },
@@ -116,10 +116,10 @@ class Providerdashboard extends StatelessWidget {
             // Shop Profile button
             ListTile(
               leading: Icon(Icons.store),
-              title: Text('Provider Profile'),
+              title: Text('Shop Profile'),
               onTap: () {
                 Navigator.push(context,
-                    MaterialPageRoute(builder: (context) => ProviderProfilePage()));
+                    MaterialPageRoute(builder: (context) => ShopProfilePage()));
                 // Handle Shop Profile button press
               },
             ),
@@ -128,7 +128,7 @@ class Providerdashboard extends StatelessWidget {
               title: Text('Feedback'),
               onTap: () {
                 Navigator.push(context,
-                    MaterialPageRoute(builder: (context) => PrviderFeedbackPage()));
+                    MaterialPageRoute(builder: (context) => ReviewsPage()));
                 // Handle Shop Profile button press
               },
             ),
@@ -136,11 +136,13 @@ class Providerdashboard extends StatelessWidget {
             const SizedBox(
               height: 25,
             ),
-
+            // Log Out button
             ElevatedButton(
               onPressed: () {
                 Navigator.push(context,
                     MaterialPageRoute(builder: (context) => HomePage()));
+
+                // Handle Log Out button press
               },
               style: ElevatedButton.styleFrom(
                 foregroundColor: const Color.fromARGB(255, 8, 0, 0),
@@ -159,14 +161,13 @@ class Providerdashboard extends StatelessWidget {
               child: ListView(
                 children: [
                   orderItem(
-                      subStatus: '2ND ORDER',
-                      orderId: '162267901',
-                      date: '12 Sept 2024, 9:31 am',
-                      ordername: 'Santize full home',
-                      statusColor: Colors.brown,
-                      customername: 'Rishaf',
-                      customermobile: '0755354023',
-                      customerLocation: 'No-2,Kinniya'),
+                    subStatus: '2ND ORDER',
+                    orderId: '162267901',
+                    date: '12 Sept 2024, 9:31 am',
+                    itemCount: 4,
+                    itemname: 'Hammer',
+                    statusColor: Colors.green,
+                  ),
                 ],
               ),
             ),
@@ -184,11 +185,8 @@ class Providerdashboard extends StatelessWidget {
       {required String subStatus,
       required String orderId,
       required String date,
-      required customername,
-      required customermobile,
-      required customerLocation,
-      required String ordername,
-
+      required int itemCount,
+      required String itemname,
       required MaterialColor statusColor}) {
     return Card(
       margin: EdgeInsets.symmetric(vertical: 10, horizontal: 15),
@@ -199,22 +197,11 @@ class Providerdashboard extends StatelessWidget {
           children: [
             Text(
               'Order ID: $orderId',
-              style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+              style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
             ),
             SizedBox(height: 5),
-            Text('Order: $ordername ',
-            style: TextStyle(
-                    color:  Colors.black,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 24.0,
-                  ),
-            ),
-            SizedBox(height: 8),
+            Text('Item: $itemname ($itemCount pcs)'),
             Text('Date: $date'),
-            Text('Customer Name: $customername'),
-            Text('Customer Location: $customerLocation'),
-            Text('Customer Mobile: $customermobile'),
-            
             SizedBox(height: 10),
             Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
