+++ FrontEnd/lib/ShopOwner/ToolsManagement.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:tap_on/AddTools.dart';
+import 'package:tap_on/ShopOwner/AddTools.dart';
 
 class MenuScreen extends StatefulWidget {
   @override
@@ -73,7 +73,7 @@ class _MenuScreenState extends State<MenuScreen> {
                   itemCount: menuItems.length,
                   itemBuilder: (context, index) {
                     var item = menuItems[index];
-                    
+
                     return Card(
                       elevation: 4,
                       child: ListTile(
@@ -89,39 +89,37 @@ class _MenuScreenState extends State<MenuScreen> {
                         subtitle: Text(
                             'Quantity: ${item['quantity']}\nLKR ${item['price']}'),
                         trailing: Row(
-                            mainAxisSize: MainAxisSize.min,
-                            children: [
+                          mainAxisSize: MainAxisSize.min,
+                          children: [
+                            Flexible(
+                              child: Text(
+                                item['available']
+                                    ? 'Available'
+                                    : 'Not Available',
+                                style: TextStyle(
+                                    color: item['available']
+                                        ? Colors.green
+                                        : Colors.red),
+                              ),
+                            ),
 
-                              Flexible(
-                                child: Text(
-                                  item['available']
-                                      ? 'Available'
-                                      : 'Not Available',
-                                  style: TextStyle(
-                                      color: item['available']
-                                          ? Colors.green
-                                          : Colors.red),
-                                ),
-                              ),
-
-                              SizedBox(width: 8),
-                              // Edit Button
-                              IconButton(
-                                icon: Icon(Icons.edit),
-                                onPressed: () {
-                                  // Open the dialog to edit item
-                                  _editItemDialog(context, item, index);
-                                },
-                              ),
-                              TextButton(
-                                onPressed: () {
-                                  _editItemDialog(context, item, index);
-                                },
-                                child: Text('Edit'),
-                              ),
-                            ],
-                          ),
-                        
+                            SizedBox(width: 8),
+                            // Edit Button
+                            IconButton(
+                              icon: Icon(Icons.edit),
+                              onPressed: () {
+                                // Open the dialog to edit item
+                                _editItemDialog(context, item, index);
+                              },
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                _editItemDialog(context, item, index);
+                              },
+                              child: Text('Edit'),
+                            ),
+                          ],
+                        ),
                       ),
                     );
                   },
