+++ FrontEnd/lib/database/config.dart
@@ -1,0 +1,3 @@
+final url = 'http://10.11.12.149:3000/';
+
+final registration = url + "registration";
