+++ FrontEnd/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:tap_on/constants.dart';
-import 'package:tap_on/laungh%20page.dart';
+import 'package:tap_on/Launch%20app.dart';
 
 //
