+++ FrontEnd/pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   cupertino_icons: ^1.0.8
   flutter_polyline_points: ^2.1.0
   google_maps_flutter: ^2.9.0
+  http: ^1.2.2
 
 dev_dependencies:
   flutter_test:
